refactor(db): split the disk format into modules
