+++ .github/workflows/test.yml
@@ -212,7 +212,10 @@ jobs:
             /zebra-state/**/constants.rs
             /zebra-state/**/finalized_state.rs
             /zebra-state/**/disk_format.rs
-            /zebra-state/**/disk_format/types.rs
+            /zebra-state/**/disk_format/block.rs
+            /zebra-state/**/disk_format/chain.rs
+            /zebra-state/**/disk_format/shielded.rs
+            /zebra-state/**/disk_format/transparent.rs
             /zebra-state/**/disk_db.rs
             /zebra-state/**/zebra_db.rs
             /zebra-state/**/zebra_db/block.rs
